+++ CHANGELOG.md
@@ -4,7 +4,11 @@
 
 ### Added
 
-- [TD-4075] Grant approver api 
+- [TD-4075] API routes for managing grant approvers:
+  - `GET /api/grant_approvers`
+  - `GET /api/grant_approvers/:id`
+  - `POST /api/grant_approvers`
+  - `DELETE /api/grant_approvers/:id`
 
 ### Changed